+++ typescript/.storybook/main.ts
@@ -13,6 +13,7 @@ const config: StorybookConfig = {
         getAbsolutePath("@storybook/addon-actions"),
         getAbsolutePath("@storybook/addon-storysource"),
         getAbsolutePath("@storybook/addon-webpack5-compiler-babel"),
+        getAbsolutePath("@storybook/addon-interactions"),
     ],
     webpackFinal: async (config) => {
         return {
@@ -86,6 +87,7 @@ const config: StorybookConfig = {
 };
 export default config;
 
+// @ts-expect-error TS7006
 function getAbsolutePath(value) {
     return dirname(require.resolve(join(value, "package.json")));
 }