+++ typescript/packages/subsurface-viewer/src/layers/grid3d/grid3dLayer.ts
@@ -48,6 +48,7 @@ export type WebWorkerParams = {
     points: Float32Array;
     polys: Uint32Array;
     properties: Float32Array | Uint16Array;
+    undefinedValue: number;
 };
 
 function GetBBox(points: Float32Array): BoundingBox3D {
@@ -199,6 +200,22 @@ export interface Grid3DLayerProps extends ExtendedLayerProps {
      */
     colorMapFunction?: colorMapFunctionType | Uint8Array;
 
+    /**
+     * Value in propertiesData indicating that the property is undefined.
+     * When propertiesData is Uint16Array the value is index in discretePropertyValueNames if provided.
+     * By default, it is assumed to be NaN for Float32 properties data and 0xFFFF for UInt16.
+     * Note: in case of floating point properties exact comparison is performed so use the same
+     * constant value both in propertiesData array and as undefinedPropertyValue.
+     */
+    undefinedPropertyValue?: number;
+
+    /**
+     * Color for the cells with undefined property value.
+     * Is not overridden by and used prior to colorMapFunction.
+     * By default, Light gray if not provided.
+     */
+    undefinedPropertyColor?: [number, number, number];
+
     /** Enable lines around cell faces.
      *  default: true.
      */
@@ -275,6 +292,7 @@ export default class Grid3DLayer extends CompositeLayer<Grid3DLayerProps> {
                 points,
                 polys,
                 properties,
+                undefinedValue: this.getUndefinedPropertyValue(),
             };
 
             pool.exec(makeFullMesh, [{ data: webworkerParams }]).then((e) => {
@@ -350,6 +368,8 @@ export default class Grid3DLayer extends CompositeLayer<Grid3DLayerProps> {
             // isFinishedLoading only in state
             return [];
         }
+        const undefinedColor = this.getUndefinedPropertyColor();
+        const undefinedValue = this.getUndefinedPropertyValue();
 
         const layer = new PrivateLayer(
             this.getSubLayerProps({
@@ -359,6 +379,8 @@ export default class Grid3DLayer extends CompositeLayer<Grid3DLayerProps> {
                 colorMapName: this.props.colorMapName,
                 colorMapRange: this.props.colorMapRange,
                 colorMapClampColor: this.props.colorMapClampColor,
+                undefinedPropertyValue: undefinedValue,
+                undefinedPropertyColor: undefinedColor,
                 colorMapFunction: this.props.colorMapFunction,
                 coloringMode: this.props.coloringMode,
                 gridLines: this.props.gridLines,
@@ -386,6 +408,20 @@ export default class Grid3DLayer extends CompositeLayer<Grid3DLayerProps> {
             default:
                 return this.state["propertyValueRange"] as [number, number];
         }
+    }
+
+    private getUndefinedPropertyValue(): number {
+        if (typeof this.props.undefinedPropertyValue === "number") {
+            return this.props.undefinedPropertyValue;
+        }
+        if (this.props.propertiesData instanceof Uint16Array) {
+            return 0xffff;
+        }
+        return Number.NaN;
+    }
+
+    private getUndefinedPropertyColor(): [number, number, number] {
+        return this.props.undefinedPropertyColor ?? [0.8, 0.8, 0.8];
     }
 }