+++ typescript/packages/well-log-viewer/src/components/ZoomSlider.stories.tsx
@@ -37,9 +37,12 @@ const stories: Meta = {
 };
 export default stories;
 
+// @ts-expect-error TS7006
 const Template = (args) => {
     const infoRef = React.useRef();
+    // @ts-expect-error TS7006
     const setInfo = function (info) {
+        // @ts-expect-error TS2339
         if (infoRef.current) infoRef.current.innerHTML = info;
     };
 
@@ -55,6 +58,8 @@ const Template = (args) => {
                     }}
                 />
             </div>
+            {/*
+             // @ts-expect-error TS2322 */}
             <div ref={infoRef}>&nbsp;</div>
         </div>
     );